Refactored frontend (#528)
* Refactoring frontend

* Fixed test

* Reduce complexity

* Code quality

* Renaming per suggestion of @tgy
