+++ neurolang/frontend/probabilistic_frontend.py
@@ -10,48 +10,48 @@ import typing
 from typing import (
     AbstractSet,
     Any,
+    Callable,
     Dict,
     Iterable,
     List,
     Optional,
     Tuple,
-    Type,
-    Callable,
+    Type
 )
 from uuid import uuid1
 
+from .. import expressions as ir
 from ..datalog.aggregation import (
     Chase,
     DatalogWithAggregationMixin,
-    TranslateToLogicWithAggregation,
+    TranslateToLogicWithAggregation
 )
 from ..datalog.constraints_representation import DatalogConstraintsProgram
 from ..datalog.ontologies_parser import OntologyParser
 from ..datalog.ontologies_rewriter import OntologyRewriter
 from ..exceptions import UnsupportedQueryError
-from .. import expressions as ir
 from ..expression_walker import ExpressionBasicEvaluator
 from ..logic import Union
 from ..probabilistic.cplogic.program import (
     CPLogicMixin,
-    TranslateProbabilisticQueryMixin,
+    TranslateProbabilisticQueryMixin
 )
 from ..probabilistic.dichotomy_theorem_based_solver import (
-    solve_succ_query as lifted_solve_succ_query,
+    solve_succ_query as lifted_solve_succ_query
 )
 from ..probabilistic.expression_processing import (
     is_probabilistic_predicate_symbol,
-    is_within_language_succ_query,
+    is_within_language_succ_query
 )
 from ..probabilistic.query_resolution import compute_probabilistic_solution
 from ..probabilistic.stratification import stratify_program
 from ..region_solver import RegionSolver
 from ..relational_algebra import (
     NamedRelationalAlgebraFrozenSet,
-    RelationalAlgebraStringExpression,
+    RelationalAlgebraStringExpression
 )
-from . import QueryBuilderDatalog
 from . import query_resolution_expressions as fe
+from .query_resolution_datalog import QueryBuilderDatalog
 
 
 class RegionFrontendCPLogicSolver(
@@ -66,7 +66,7 @@ class RegionFrontendCPLogicSolver(
     pass
 
 
-class ProbabilisticFrontend(QueryBuilderDatalog):
+class NeurolangPDL(QueryBuilderDatalog):
     """
     Complements QueryBuilderDatalog class with probabilistic capabilities
     1- add extensional probabilistic facts and choices
@@ -165,10 +165,7 @@ class ProbabilisticFrontend(QueryBuilderDatalog):
 
     def _execute_query(
         self,
-        head: typing.Union[
-            fe.Symbol,
-            Tuple[fe.Expression, ...],
-        ],
+        head: typing.Union[fe.Symbol, Tuple[fe.Expression, ...]],
         predicate: fe.Expression,
     ) -> Tuple[AbstractSet, Optional[ir.Symbol]]:
         """
@@ -263,9 +260,7 @@ class ProbabilisticFrontend(QueryBuilderDatalog):
         )
         return solution, functor_orig
 
-    def solve_all(
-        self,
-    ) -> Dict[str, NamedRelationalAlgebraFrozenSet]:
+    def solve_all(self,) -> Dict[str, NamedRelationalAlgebraFrozenSet]:
         """
         Returns a dictionary of "predicate_name": "Content"
         for all elements in the solution of the Datalog program.
@@ -361,8 +356,7 @@ class ProbabilisticFrontend(QueryBuilderDatalog):
         solver = RegionFrontendCPLogicSolver()
         for psymb, relation in solution.items():
             solver.add_extensional_predicate_from_tuples(
-                psymb,
-                relation.value,
+                psymb, relation.value,
             )
         for builtin_symb in self.program_ir.builtins():
             solver.symbol_table[builtin_symb] = self.program_ir.symbol_table[
@@ -607,3 +601,6 @@ class ProbabilisticFrontend(QueryBuilderDatalog):
         ra_set = ra_set.extended_projection(projections)
         self.program_ir.add_probabilistic_choice_from_tuples(symbol, ra_set)
         return fe.Symbol(self, name)
+
+
+ProbabilisticFrontend = NeurolangPDL
